Lesson listing and datatable function added to course,instructor,category

+++ database/migrations/2022_03_07_035635_create_courses_table.php
@@ -34,9 +34,9 @@ class CreateCoursesTable extends Migration
             $table->date('enroll_close_date')->nullable();
 
             $table->tinyInteger('published')->default(0)->nullable();
-            $table->integer('featured')->default(0)->nullable();
-            $table->integer('trending')->default(0)->nullable();
-            $table->integer('popular')->default(0)->nullable();
+            $table->tinyInteger('featured')->default(0)->nullable();
+            $table->tinyInteger('trending')->default(0)->nullable();
+            $table->tinyInteger('popular')->default(0)->nullable();
 
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
             $table->foreign('instructor_id')->references('id')->on('instructors')->onDelete('cascade');
